move common code to separate module common.py
