+++ examples/python/example_without_zone.py
@@ -13,17 +13,10 @@
 # limitations under the License.
 # coding: utf-8
 import argparse
-import os
 
 from acrawriter.sqlalchemy import AcraBinary
-from sqlalchemy import (Table, Column, Integer, MetaData, create_engine,
-                        select, Text)
-
-
-def get_default(name, value):
-    """return value from environment variables with name EXAMPLE_<name>
-    or value"""
-    return os.environ.get('EXAMPLE_{}'.format(name.upper()), value)
+from sqlalchemy import (Table, Column, Integer, MetaData, select, Text)
+from common import get_engine, get_default, register_common_cli_params
 
 
 def print_data(connection):
@@ -43,70 +36,16 @@ def write_data(data, connection):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
+    register_common_cli_params(parser)
     parser.add_argument(
         '--public_key', type=str, default=get_default('public_key', ''),
         help='path to acra-server public key '
              '(for example .acrakeys/<client_id>_storage.pub)')
-    parser.add_argument('--db_name', type=str,
-                        default=get_default('db_name', 'acra'),
-                        help='Database name')
-    parser.add_argument('--db_user', type=str,
-                        default=get_default('db_user', 'test'),
-                        help='Database user')
-    parser.add_argument('--db_password', type=str,
-                        default=get_default('db_password', 'test'),
-                        help='Database user\'s password')
-    parser.add_argument('--port', type=int,
-                        default=get_default('port', 9494),
-                        help='Port of database or AcraConnector')
-    parser.add_argument('--host', type=str,
-                        default=get_default('host', 'localhost'),
-                        help='Host of database or AcraConnector')
     parser.add_argument('--data', type=str,
                         default=get_default('data', ''),
                         help='data to save in ascii. default random data')
-    parser.add_argument('--ssl_mode', action='store_true',
-                        default=get_default('ssl_mode', False),
-                        help='SSL connection mode')
-    parser.add_argument('--tls_root_cert', action='store_true',
-                        default=get_default('tls_root_cert', False),
-                        help='Path to root certificate used in TLS connection')
-    parser.add_argument('--tls_key', action='store_true',
-                        default=get_default('tls_key', False),
-                        help='Path to client TLS key used in TLS connection')
-    parser.add_argument('--tls_cert', action='store_true',
-                        default=get_default('tls_cert', False),
-                        help='Path to client TLS certificate used in TLS connection')
-    parser.add_argument('--print', action='store_true',
-                        default=get_default('print', False),
-                        help='Print data (use --zone_id to set specific ZoneId '
-                             'which will be used to fetch data)')
-    parser.add_argument('-v', '--verbose', dest='verbose', action='store_true',
-                        default=get_default('verbose', False), help='verbose')
-    parser.add_argument('--postgresql', action='store_true',
-                        default=get_default('postgresql', False),
-                        help="Use postgresql driver (default if nothing else "
-                             "set)")
-    parser.add_argument('--mysql', action='store_true',
-                        default=get_default('mysql', False),
-                        help="Use mysql driver")
     args = parser.parse_args()
 
-    # default driver
-    driver = 'postgresql'
-    ssl_args = {
-        'sslmode': args.ssl_mode,
-        'sslrootcert': args.tls_root_cert,
-        'sslkey': args.tls_key,
-        'sslcert': args.tls_cert,
-    }
-    if args.mysql:
-        driver = 'mysql+pymysql'
-        ssl_args = {
-            'ssl_ca': args.tls_root_cert,
-            'ssl_cert': args.tls_cert,
-            'ssl_key': args.tls_key
-        }
 
     metadata = MetaData()
     # here we load public key for AcraStructs
@@ -120,16 +59,13 @@ if __name__ == '__main__':
         Column('data', AcraBinary(key), nullable=False),
         Column('raw_data', Text, nullable=False))
 
-    engine = create_engine(
-        '{}://{}:{}@{}:{}/{}'.format(
-            driver, args.db_user, args.db_password, args.host, args.port,
-            args.db_name),
-        connect_args=ssl_args,
-        echo=bool(args.verbose))
+    engine = get_engine(
+        db_host=args.host, db_port=args.port, db_user=args.db_user, db_password=args.db_password,
+        db_name=args.db_name, is_mysql=args.mysql, is_postgresql=args.postgresql,
+        tls_ca=args.tls_root_cert, tls_key=args.tls_key, tls_crt=args.tls_cert,
+        sslmode=args.ssl_mode, verbose=args.verbose)
     metadata.create_all(engine)
     connection = engine.connect()
-
-    print('DB driver: {}'.format(driver))
 
     if args.print:
         print_data(connection)
